requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
attrs==18.1.0
cffi==1.11.5
chardet==3.0.4
discord.py==0.16.12
idna==2.7
idna-ssl==1.1.0
multidict==4.3.1
pycparser==2.18
PyNaCl==1.2.1
PyYAML==3.13
six==1.11.0
websockets==3.4
yarl==1.2.6
youtube-dl==2018.7.21
